% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_biomass}
\alias{add_biomass}
\title{Given a table of observations as biomass as computed in Sorochan et al 2019}
\usage{
add_biomass(
  x = read_staged(form = "sf"),
  region = c("GoM-GBK, SS", "All regions", "GSL, NLS", "southwest GSL")[1],
  params = Sorochan(taxa = "C. finmarchicus")
)
}
\arguments{
\item{x}{tibble or sf, staged data from ecomon}

\item{region}{character, the region for application - see \code{\link[Sorochan]}}

\item{params}{tibble, parameters as per \code{\link{Sorochan}}}
}
\value{
the input with an added column called "biomass"
}
\description{
Given a table of observations as biomass as computed in Sorochan et al 2019
}
