% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmp.R
\name{correct_azmp}
\alias{correct_azmp}
\title{Given staged ecomon data correct to allow merging with AZMP data}
\usage{
correct_azmp(
  x = read_staged(species = "calfin", form = "tibble"),
  model = read_azmp_model(filename = "Cfin_CIV_CVI"),
  stages = c("c4_m2", "c5_m2", "c6_m2"),
  max_depth = 500,
  handle_deep = c("drop", "clip")[1],
  sorochan = TRUE,
  max_percent_z = 95,
  min_above_bottom = 15,
  min_station_depth = 40
)
}
\arguments{
\item{x}{data frame of staged data}

\item{model}{gam model from the mgcv package}

\item{stages}{str, the stages to combine.  "all" to combine all}

\item{max_depth}{numbers, depths above this must be manages as per \link{handle_deep}}

\item{handle_deep}{str, one of 'drop', 'clip' or 'skip' (to skip managing deeper stations)
\code{drop} will drop stations above \code{max_depth}, \code{clip} will clip stations to \code{max_depth}
and any other value, like skip, will skip this drop/clip step.}

\item{sorochan}{logical, if TRUE then perform filtering that includes
stations where the tow depth is less that \code{max_percent_z} of station depth,
AND excludes stations that are within \code{min_above_bottom}
AND have a minimum station depth of at least \code{min_station_depth}}

\item{max_percent_z}{num maximum percent of tow depth relative to station depth}

\item{min_above_bottom}{num, meters of tow depth above station depth}

\item{min_station_depth}{num, meters of minimum station depth}
}
\value{
data frame input including computed abundance estimates
}
\description{
Given staged ecomon data correct to allow merging with AZMP data
}
