% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ecomon_species}
\alias{ecomon_species}
\title{Retrieve a table of known ecomon species and metrics}
\usage{
ecomon_species(x = ecomon_cols())
}
\arguments{
\item{x}{named character vector as produced by \code{ecomon_cols()}}
}
\value{
table of \code{name}, \code{metric}
}
\description{
Retrieve a table of known ecomon species and metrics
}
