% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{ecomon_url}
\alias{ecomon_url}
\title{Retrieve a NCEI url to a dataset}
\usage{
ecomon_url(
  id = "0187513",
  base_url = "https://www.ncei.noaa.gov/archive/accession/download"
)
}
\arguments{
\item{id}{char, a dataset id like '0187513'}

\item{base_url}{char, the root path to NCEI datasets}
}
\value{
chaacter URL
}
\description{
Retrieve a NCEI url to a dataset
}
