% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_data_path}
\alias{get_data_path}
\title{Retrieve the user specified data path}
\usage{
get_data_path(..., root = readLines("~/.ecomon"))
}
\arguments{
\item{...}{elements for \code{\link[base]{file.path}}}

\item{root}{char, the root path directory}
}
\value{
character, the path to the data
}
\description{
Retrieve the user specified data path
}
