% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{list_data}
\alias{list_data}
\title{Retrieve the path to one or more ecomon datasets}
\usage{
list_data(
  id = "0187513",
  most_recent = TRUE,
  version = c("2.2", "recent")[2],
  dataset = "0-data",
  path = get_data_path()
)
}
\arguments{
\item{id}{character, the accession id of the data set ala "0187513"}

\item{version}{character, defaults to "recent" but could be "1.1", "2.2", etc}

\item{dataset}{character, one "0-data" (default) or "1-data"}

\item{path}{character the root path}
}
\value{
character vector of filenames
}
\description{
Retrieve the path to one or more ecomon datasets
}
