% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmp.R
\name{read_azmp_example}
\alias{read_azmp_example}
\title{Read AZMP correction data example}
\usage{
read_azmp_example(filename = "dattax_Cfin_IV-VI.csv.gz")
}
\arguments{
\item{filename}{str the name of the file to read}
}
\value{
table of example data
}
\description{
Read AZMP correction data example
}
