% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmp.R
\name{read_azmp_model}
\alias{read_azmp_model}
\title{Read an AZMP correction model}
\usage{
read_azmp_model(filename = c("Cfin_CIV_CVI", "Cfin_CV_CVI")[1])
}
\arguments{
\item{filename}{str, one of 'Cfin_CIV_CVI' or 'Cfin_CV_CVI'}
}
\value{
\code{\link[mgcv:gamm]{mgcv::gamm()}} model
}
\description{
Read an AZMP correction model
}
