% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_cross}
\alias{read_cross}
\title{Read the special file provided to C Ross}
\usage{
read_cross(
  filename = get_data_path("EcoMon_CalfinStage_Thru_12_30_2019_10m2.csv"),
  form = c("tibble", "sf")[1]
)
}
\arguments{
\item{filename}{file specification}

\item{form}{character either 'tibble' or 'sf'}
}
\value{
tibble or sf Points object
}
\description{
Read the special file provided to C Ross
}
