% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ctd}
\alias{read_ctd}
\title{Read NOAA CTD table provided by personal communication by Chris Melrose to Chris Orphanides}
\usage{
read_ctd(
  path = get_data_path("ctd"),
  match_cruise = NULL,
  form = c("tibble", "sf")[1]
)
}
\arguments{
\item{path}{character, the path to the data}

\item{match_cruise}{NULL or data frame (tibble), if not \code{NULL}
then filter the output to match the cruise IDs in the match_cruise
object.}

\item{form}{character, one of 'tibble' or 'sf'}

\item{simplify}{logical, if TRUE simply the dataset}
}
\value{
'tibble' or 'sf' object
}
\description{
Read NOAA CTD table provided by personal communication by Chris Melrose to Chris Orphanides
}
