% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ecomon}
\alias{read_ecomon}
\title{Read a ecomon data - trim to bare necessities}
\usage{
read_ecomon(
  filename = list_data(id = "0187513"),
  simplify = TRUE,
  select_vars = c("cruise_name", "station", "zoo_gear", "ich_gear", "lat", "lon",
    "date", "time", "depth", "sfc_temp", "sfc_salt", "btm_temp", "btm_salt",
    "volume_1m2"),
  form = c("tibble", "sf")[1]
)
}
\arguments{
\item{filename}{one or more filenames.  If multiples, the
individual tables are row bound into one.}

\item{simplify}{logical if TRUE trim down to a simple dataset}

\item{select_vars}{character, the columns to select if simplyfing.  Ignored
unless \code{simplify = TRUE}.}

\item{form}{character either 'tibble' or 'sf'}
}
\value{
tibble or sf Points object
}
\description{
Read a ecomon data - trim to bare necessities
}
