% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ecomon_one}
\alias{read_ecomon_one}
\title{Read a a single ecomon data file}
\usage{
read_ecomon_one(
  filename,
  skip = 0,
  col_types = paste(ecomon_cols(), collapse = ""),
  to_lower_case = TRUE
)
}
\arguments{
\item{filename}{the name of the file}

\item{skip}{the number of header lines to skip}

\item{col_types}{a vector or compact character string used to declare input column types}

\item{to_lower_case}{logocal, if TRUE make the columns names lower case (this
make v3.11 the same as earlier versions)}
}
\value{
tibble
}
\description{
Read a a single ecomon data file
}
