% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_staged}
\alias{read_staged}
\title{Read the staged data provided via 2022-03-03 personal communication (to modeling group)
by Harvey Walsh}
\usage{
read_staged(species = "calfin", form = c("tibble", "sf")[1], ...)
}
\arguments{
\item{species}{character, the species to read}

\item{form}{character either 'tibble' or 'sf'}

\item{...}{other arguments for \code{\link{staged_filename}}}
}
\value{
tibble or sf Points object
}
\description{
Read the staged data provided via 2022-03-03 personal communication (to modeling group)
by Harvey Walsh
}
