% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoop.R
\name{read_zoop}
\alias{read_zoop}
\title{Read the Zoo-Ichy Plankton data sheet.}
\usage{
read_zoop(what = c("m2", "m3")[2], filename = NULL)
}
\arguments{
\item{what}{character, either 'm2' or 'm3', ignored if \code{filename} is not
\code{NULL}}

\item{filename}{character or NULL.  If NULL then the package's data is read.
if not NULL then the specified file is read.}
}
\value{
a data.frame or NULL
}
\description{
Read the Zoo-Ichy Plankton data sheet.
}

