% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{scale_ecomon}
\alias{scale_ecomon}
\title{Scales ecomon units}
\usage{
scale_ecomon(
  x = read_ecomon(simplify = FALSE),
  scale = c(`_10m2` = 10, `_100m3` = 100),
  rename = c(`_10m2` = "_m2", `_100m3` = "_m3")
)
}
\arguments{
\item{x}{df, table of ecomon data}

\item{scale}{named vector, desired scale factor named by key pattern}

\item{rename}{named vector, desired new pattern named by old pattern}
}
\value{
df, ecomon dataset scaled as desired
}
\description{
Scales ecomon units
}
