% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{show_docs}
\alias{show_docs}
\title{Show the documentation (aka README) associated with one of the
 included datasets.}
\usage{
show_docs(what = "zoop")
}
\arguments{
\item{what}{charact name of the dataset, currelty only 'zoop'}
}
\description{
Show the documentation (aka README) associated with one of the
 included datasets.
}

