% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{staged_filename}
\alias{staged_filename}
\title{Given a species, retirve the name of the staged species file}
\usage{
staged_filename(
  species = "calfin",
  path = get_data_path("staged"),
  must_exist = TRUE
)
}
\arguments{
\item{species}{char, the species to read}

\item{path}{char, the path to the species staged data}

\item{must_exist}{logical, if TRUE test for the existence of the file, throw an
error if not found}
}
\value{
character, file specification
}
\description{
Given a species, retirve the name of the staged species file
}
