% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{staged_match_ctd}
\alias{staged_match_ctd}
\title{Given a cruise data frame of staged data with \code{cruise_name}
generate a matching subset of the ctd data}
\usage{
staged_match_ctd(
  x = dplyr::filter(read_staged(), .data$cruise_name \%in\% c("PE9602", "HB1906")),
  ctd = read_ctd(),
  form = c("list", "tibble")[1]
)
}
\arguments{
\item{x}{tibble of staged cruise data}

\item{ctd}{tibble of ctd data}

\item{form}{character, the form as a named 'list' (each element one data frame) or 
a 'tibble' where all results are bound in one table}
}
\value{
tibble of ctd data, possibly empty, for matches by cruise 
  or a list of the same
}
\description{
Given a cruise data frame of staged data with \code{cruise_name}
generate a matching subset of the ctd data
}
