% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoop.R
\name{Zoop}
\alias{Zoop}
\title{Create an instance of a ZoopRefClass object}
\usage{
Zoop(...)
}
\arguments{
\item{...}{further arguments for \code{\link{read_zoop}}}
}
\value{
an instance of ZoopRefClass
}
\description{
Create an instance of a ZoopRefClass object
}

