% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoop.R
\docType{class}
\name{ZoopRefClass-class}
\alias{ZoopRefClass}
\alias{ZoopRefClass-class}
\title{A simple container class for the Zoop data set.}
\description{
A simple container class for the Zoop data set.
}
\section{Fields}{

\describe{
\item{\code{version}}{character, the data version if known}

\item{\code{data}}{a data.frame or NULL}
}}

