% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zoop.R
\name{ZoopRefClass_get}
\alias{ZoopRefClass_get}
\title{Get one or more records}
\arguments{
\item{species}{character one or more species names to retrieve or 'all'
NOTE: if you try c("salps", "all") you will get "all"}

\item{date_filter}{POSIXct or Date, two element vector populated according to
By default NULL.  Records within the date_filter[1] <=  >= date_filter[2]
are returned.}

\item{projargs}{character, by default "+proj=longlat +datum=WGS84"}

\item{form}{character either 'data.frame' or 'SpatialPointsDataFrame'}

\item{bb}{a 4 element vector of [xmin, xmax, ymin, ymax] or NULL bounding box.
Data are clipped to this bbox.}
}
\value{
data.frame, SpatialPointsDataFrame or NULL invisibly
}
\description{
Get one or more records
}

