% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoop.R
\name{ZoopRefClass_retain}
\alias{ZoopRefClass_retain}
\title{Retain one or more columns - others are removed}
\arguments{
\item{name}{character vector of one or more columns to retain}
}
\description{
Retain one or more columns - others are removed
}

