% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomon-package.R
\docType{package}
\name{ecomon}
\alias{ecomon}
\title{ecomon - manage local data of NMFS-NEFSC ecomon database}
\description{
Functions for reading [EcoMon](https://www.fisheries.noaa.gov/about/northeast-fisheries-science-center)
}
