% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ecomon_cols}
\alias{ecomon_cols}
\title{Retrieve a named vector of column names and types for ecomon}
\usage{
ecomon_cols(var_names = character())
}
\arguments{
\item{var_names}{character, identifies variable names to keep, all other marked to skip}
}
\value{
listing of column type definitions
}
\description{
Retrieve a named vector of column names and types for ecomon
}
