% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_ecomon}
\alias{fetch_ecomon}
\title{Fecth EcoMon data from NCEI}
\usage{
fetch_ecomon(
  id = "0187513",
  dest = get_data_path(),
  decompress = TRUE,
  cleanup = decompress
)
}
\arguments{
\item{id}{char, a dataset id like '0187513'}

\item{dest}{character, the path to save the data to}

\item{decompress}{logical, if TRUE unpack the downloaded file}

\item{cleanup}{logical, if TRUE remove the downloaded zip file. Ignored unless
\code{decompress} is \code{TRUE}}
}
\value{
the result of \code{\link[utils]{download.file}}
}
\description{
Fecth EcoMon data from NCEI
}
