% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ecomon_one}
\alias{read_ecomon_one}
\title{Read a a single ecomon data file}
\usage{
read_ecomon_one(
  filename,
  skip = 0,
  col_types = paste(ecomon_cols(), collapse = "")
)
}
\arguments{
\item{filename}{the name of the file}

\item{skip}{the number of header lines to skip}

\item{col_types}{a vector or compact character string used to declare input column types}
}
\value{
tibble
}
\description{
Read a a single ecomon data file
}
