% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{Sorochan}
\alias{Sorochan}
\title{Read data from the Sorochan paper}
\usage{
Sorochan(what = "table_1")
}
\arguments{
\item{what}{character, the name of the item to read}
}
\value{
tibble
}
\description{
Read data from the Sorochan paper
}
